Move enqueue to queue class (untested
