+++ src/client/client.ts
@@ -1,11 +1,11 @@
 import { DLQ } from "./dlq";
 import { HttpClient, Requester, RetryConfig } from "./http";
 import { Messages } from "./messages";
-import { Queues } from "./queue";
+import { Queue } from "./queue";
 import { Schedules } from "./schedules";
 import { Topics } from "./topics";
 import { Event } from "./types";
-import { prefixHeaders } from "./utils";
+import { prefixHeaders, processHeaders } from "./utils";
 type ClientConfig = {
   /**
    * Url of the qstash api server.
@@ -132,21 +132,21 @@ export type PublishRequest<TBody = BodyInit> = {
    */
   method?: "GET" | "POST" | "PUT" | "DELETE" | "PATCH";
 } & (
-  | {
+    | {
       /**
        * The url where the message should be sent to.
        */
       url: string;
       topic?: never;
     }
-  | {
+    | {
       url?: never;
       /**
        * The url where the message should be sent to.
        */
       topic: string;
     }
-);
+  );
 
 export type PublishJsonRequest = Omit<PublishRequest, "body"> & {
   /**
@@ -165,10 +165,9 @@ export type GetEventsResponse = {
   events: Event[];
 };
 
-export type EnqueueRequest = {
-  publishRequest: PublishRequest;
+export type QueueRequest = {
   queueName: string;
-};
+}
 
 export class Client {
   public http: Requester;
@@ -224,50 +223,14 @@ export class Client {
    * 
    * Create, read, update or delete queues.
    */
-  public get queues(): Queues {
-    return new Queues(this.http);
-  }
-
-  private processHeaders(req: PublishRequest) {
-    const headers = prefixHeaders(new Headers(req.headers));
-
-    headers.set("Upstash-Method", req.method ?? "POST");
-
-    if (typeof req.delay !== "undefined") {
-      headers.set("Upstash-Delay", `${req.delay.toFixed()}s`);
-    }
-
-    if (typeof req.notBefore !== "undefined") {
-      headers.set("Upstash-Not-Before", req.notBefore.toFixed());
-    }
-
-    if (typeof req.deduplicationId !== "undefined") {
-      headers.set("Upstash-Deduplication-Id", req.deduplicationId);
-    }
-
-    if (typeof req.contentBasedDeduplication !== "undefined") {
-      headers.set("Upstash-Content-Based-Deduplication", "true");
-    }
-
-    if (typeof req.retries !== "undefined") {
-      headers.set("Upstash-Retries", req.retries.toFixed());
-    }
-
-    if (typeof req.callback !== "undefined") {
-      headers.set("Upstash-Callback", req.callback);
-    }
-
-    if (typeof req.failureCallback !== "undefined") {
-      headers.set("Upstash-Failure-Callback", req.failureCallback);
-    }
-
-    return headers;
+  public queue(req: QueueRequest): Queue {
+    return new Queue(this.http, req.queueName);
   }
 
   public async publish<TRequest extends PublishRequest>(
     req: TRequest
   ): Promise<PublishResponse<TRequest>> {
-    const headers = this.processHeaders(req);
+    const headers = processHeaders(req);
     const res = await this.http.request<PublishResponse<TRequest>>({
       path: ["v2", "publish", req.url ?? req.topic],
       body: req.body,
@@ -305,7 +268,7 @@ export class Client {
   ): Promise<PublishResponse<PublishRequest>[]> {
     const messages = [];
     for (const message of req) {
-      const headers = this.processHeaders(message);
+      const headers = processHeaders(message);
       const headerEntries = Object.fromEntries(headers.entries());
 
       messages.push({
@@ -347,44 +310,6 @@ export class Client {
     //  we can safely cast the request to `PublishRequest`
     const res = await this.batch(req as PublishRequest[]);
     return res as PublishResponse<TRequest>[];
-  }
-
-
-  /**
-   * Enqueue a message to a queue.
-   */
-  public async enqueue(req: EnqueueRequest): Promise<PublishResponse<PublishRequest>> {
-    const headers = this.processHeaders(req.publishRequest);
-    const destination = req.publishRequest.url ?? req.publishRequest.topic;
-    const res = await this.http.request<PublishResponse<PublishRequest>>({
-      path: ["v2", "enqueue", req.queueName, destination],
-      body: req.publishRequest.body,
-      headers,
-      method: "POST",
-    });
-
-    return res;
-  }
-
-  /**
-   * Enqueue a message to a queue, serializing the body to JSON.
-   */
-  public async enqueueJSON<TBody = unknown>(
-    req: EnqueueRequest & { publishRequest: PublishRequest<TBody> }
-  ): Promise<PublishResponse<PublishRequest<TBody>>> {
-    const headers = prefixHeaders(new Headers(req.publishRequest.headers));
-    headers.set("Content-Type", "application/json");
-
-    const res = await this.enqueue({
-      ...req,
-      publishRequest: {
-        ...req.publishRequest,
-        headers,
-        body: JSON.stringify(req.publishRequest.body),
-      },
-    });
-
-    return res;
   }
