+++ src/client/queue.ts
@@ -1,6 +1,8 @@
-import { Requester } from "./http";
+import type { PublishRequest, PublishResponse } from "./client";
+import type { Requester } from "./http";
+import { prefixHeaders, processHeaders } from "./utils";
 
-export type Queue = {
+export type TQueue = {
   createdAt: number,
   updatedAt: number,
   name: string,
@@ -9,59 +11,106 @@ export type Queue = {
 }
 
 export type UpsertQueueRequest = {
-  queueName: string,
   parallelism: number,
 }
 
-export class Queues {
-  private readonly http: Requester;
+export type EnqueueRequest = {
+  publishRequest: PublishRequest;
+  queueName: string;
+};
 
-  constructor(http: Requester) {
+export class Queue {
+  private readonly http: Requester;
+  private readonly queueName: string;
+
+  constructor(http: Requester, queueName: string) {
     this.http = http;
+    this.queueName = queueName;
   }
 
   /**
-   * Create or update a queue
+   * Create or update the queue
    */
   public async upsert(req: UpsertQueueRequest): Promise<void> {
+    const body = {
+      queueName: this.queueName,
+      parallelism: req.parallelism
+    }
+
     await this.http.request({
       method: "POST",
       path: ["v2", "queues"],
       headers: {
         "Content-Type": "application/json",
       },
-      body: JSON.stringify(req),
+      body: JSON.stringify(body),
     })
   }
 
   /**
-   * Get a queue
+   * Get the queue details
    */
-  public async get(queueName: string): Promise<Queue> {
-    return await this.http.request<Queue>({
+  public async get(): Promise<TQueue> {
+    return await this.http.request<TQueue>({
       method: "GET",
-      path: ["v2", "queues", queueName],
+      path: ["v2", "queues", this.queueName],
     })
   }
 
   /**
    * List queues
    */
-  public async list(): Promise<Queue[]> {
-    return await this.http.request<Queue[]>({
+  public async list(): Promise<TQueue[]> {
+    return await this.http.request<TQueue[]>({
       method: "GET",
       path: ["v2", "queues"],
     })
   }
 
   /**
-   * Delete a queue
+   * Delete the queue
    */
-  public async delete(queueName: string): Promise<void> {
+  public async delete(): Promise<void> {
     await this.http.request({
       method: "DELETE",
-      path: ["v2", "queues", queueName],
+      path: ["v2", "queues", this.queueName],
     })
   }
 
+  /**
+   * Enqueue a message to a queue.
+   */
+  public async enqueue(req: EnqueueRequest): Promise<PublishResponse<PublishRequest>> {
+    const headers = processHeaders(req.publishRequest);
+    const destination = req.publishRequest.url ?? req.publishRequest.topic;
+    const res = await this.http.request<PublishResponse<PublishRequest>>({
+      path: ["v2", "enqueue", req.queueName, destination],
+      body: req.publishRequest.body,
+      headers,
+      method: "POST",
+    });
+
+    return res;
+  }
+
+  /**
+   * Enqueue a message to a queue, serializing the body to JSON.
+   */
+  public async enqueueJSON<TBody = unknown>(
+    req: EnqueueRequest & { publishRequest: PublishRequest<TBody> }
+  ): Promise<PublishResponse<PublishRequest<TBody>>> {
+    const headers = prefixHeaders(new Headers(req.publishRequest.headers));
+    headers.set("Content-Type", "application/json");
+
+    const res = await this.enqueue({
+      ...req,
+      publishRequest: {
+        ...req.publishRequest,
+        headers,
+        body: JSON.stringify(req.publishRequest.body),
+      },
+    });
+
+    return res;
+  }
 }
